Separated model into several files (model.py -> model.py, df_maker.py. dictionary_pickler.py).
